server/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-cors-headers==4.2.0
django-pandas==0.6.6
djangorestframework==3.14.0
idna==3.4
mysqlclient==2.2.0
numpy==1.25.2
pandas==2.0.3
python-dateutil==2.8.2
pytz==2023.3
regex==2023.8.8
requests==2.31.0
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
